Move KeplerQualityFlags to targetpixelfile and use its recommended bitmask
