+++ pyke/tests/test_utils.py
@@ -1,9 +1,7 @@
-import pytest
 import argparse
 
 from ..utils import PyKEArgumentHelpFormatter
 from ..utils import module_output_to_channel, channel_to_module_output
-from ..utils import KeplerQualityFlags
 
 
 def test_PyKEArgumentHelpFormatter():
@@ -37,13 +35,3 @@ def test_module_output_to_channel():
     assert module_output_to_channel(13, 2) == 42
     assert module_output_to_channel(24, 4) == 84
     assert module_output_to_channel(11, 1) == 33
-
-
-def test_parse_kepler_quality_flags():
-    flags = list(KeplerQualityFlags.flags.items())
-    # Can we recover each individual quality flag?
-    for key, value in flags:
-        assert KeplerQualityFlags.parse(key)[0] == value
-    # Can we recover combinations of flags?
-    assert KeplerQualityFlags.parse(flags[5][0] + flags[7][0]) == [flags[5][1], flags[7][1]]
-    assert KeplerQualityFlags.parse(flags[3][0] + flags[4][0] + flags[5][0]) == [flags[3][1], flags[4][1], flags[5][1]]
